+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/window/descriptor/decorator/ExternallyReferencedDocumentDecorator.java
@@ -37,8 +37,6 @@ public class ExternallyReferencedDocumentDecorator implements IDocumentDecorator
 {
 	private static final AdMessageKey EXTERNAL_REFERENCE_READ_ONLY_IN_METASFRESH_ERROR = AdMessageKey.of("CannotDeleteExternalReferenceReadOnlyInMetasfresh");
 
-	private static final AdMessageKey EXTERNAL_REFERENCE_READ_ONLY_IN_METASFRESH_REASON = AdMessageKey.of("ExternalReferenceReadOnlyInMetasfreshReason");
-
 	private final IMsgBL msgBL = Services.get(IMsgBL.class);
 
 	private final ExternalReferenceRepository externalReferenceRepository;
@@ -52,14 +50,6 @@ public class ExternallyReferencedDocumentDecorator implements IDocumentDecorator
 	public boolean isReadOnly(@NonNull final TableRecordReference recordReference)
 	{
 		return externalReferenceRepository.isReadOnlyInMetasfresh(recordReference);
-	}
-
-	@Override
-	public BooleanWithReason getReadOnlyWithReason(@NonNull final TableRecordReference recordReference)
-	{
-		return isReadOnly(recordReference) ?
-				BooleanWithReason.trueBecause(msgBL.getMsg(EXTERNAL_REFERENCE_READ_ONLY_IN_METASFRESH_REASON, null)) :
-				BooleanWithReason.FALSE;
 	}
 
 	@Override
