+++ misc/services/camel/de-metas-camel-externalsystems/de-metas-camel-sap-file-import/src/test/resources/application.properties
@@ -1,6 +1,6 @@
 #
 # %L
-# de-metas-camel-sap
+# de-metas-camel-sap-file-import
 # %%
 # Copyright (C) 2022 metas GmbH
 # %%
