+++ misc/services/camel/de-metas-camel-externalsystems/de-metas-camel-sap-file-import/src/test/resources/de/metas/camel/externalsystems/sap/product/0_JsonStopExternalSystemRequest.json
@@ -15,6 +15,6 @@
     "SFTP_Target_Directory" : "doesntmatter",
     "SFTPProcessedDirectory": "doesntmatter",
     "SFTPErroredDirectory": "doesntmatter",
-    "SFTPPollingFrequencyInMs": 1000
+    "SFTPPollingFrequencyInMs": "1000"
   }
 }
